+++ lakin.5/oss.h
@@ -31,7 +31,7 @@ static int allocatedProcs = 0;
 static int activeProcs = 0;
 //static int msg_id = -1;
 //int g_bitVector = 0;
-char logbuf[200];
+// char logbuf[200];
 char indBuf[2];
 pid_t pid;
 sem_t *semaphore;
@@ -46,7 +46,7 @@ void createMessageQueue();
 void launchNewProc();
 void ossClock();
 void updateClock(int, int);
-void logger(const char *);
+// void logger(const char *);
 void deinitSharedMemory();
 void setBit(int);
 bool bitIsSet(int);
